// src/seaportEx/types.ts
export type OfferItem = {
    itemType: ItemType;
    token: string;
    identifierOrCriteria: string;
    startAmount: string;
    endAmount: string;
};

export type ConsiderationItem = {
    itemType: ItemType;
    token: string;
    identifierOrCriteria: string;
    startAmount: string;
    endAmount: string;
    recipient: string;
};

export type Item = OfferItem | ConsiderationItem;

export type OrderParameters = {
    offerer: string;
    zone: string;
    orderType: OrderType;
    startTime: BigNumberish;
    endTime: BigNumberish;
    zoneHash: string;
    salt: string;
    offer: OfferItem[];
    consideration: ConsiderationItem[];
    totalOriginalConsiderationItems: BigNumberish;
    conduitKey: string;
};

export type OrderComponents = OrderParameters & { counter: number };

export type Order = {
    parameters: OrderParameters;
    signature: string;
};
// ...
export interface OrdersQueryParams {
    token_ids: string[]
    asset_contract_address: string
    payment_token_address?: string
    include_bundled?: boolean
    maker?: string
    taker?: string
    side?: number
    owner?: string
    order_by?: string
    limit?: number
    offset?: number
}

export interface AssetsQueryParams {
    assets?: {
        asset_contract_addresses: string
        token_ids?: string
    }[],
    owner?: string
    limit?: number
    include_orders?: boolean
}

export interface FeesInfo {
    royaltyFeeAddress: string
    royaltyFeePoint: number
    protocolFeePoint?: number
    protocolFeeAddress?: string
}

export interface AssetCollection extends FeesInfo {
    name: string
    symbol: string
    address?: string
    token_id?: string
    schema_name?: string
    nft_version?: string
    created_date?: string
    sell_orders?: any
}
// ...
export interface ECSignature {
    v: number
    r: string
    s: string
}

export interface OrderJSON extends Partial<ECSignature> {
    exchange: string
    maker: string
    taker: string
    makerRelayerFee: string
    takerRelayerFee: string
    makerProtocolFee: string
    takerProtocolFee: string
    makerReferrerFee: string
    feeRecipient: string
    feeMethod: number
    side: number
    saleKind: number
    target: string
    howToCall: number
    calldata: string
    replacementPattern: string
    staticTarget: string
    staticExtradata: string
    paymentToken: string
    quantity: string
    basePrice: string
    englishAuctionReservePrice: string | undefined
    extra: string

    // createdTime is undefined when order hasn't been posted yet
    // createdTime?: number | string
    listingTime: number | string
    expirationTime: number | string

    salt: string

    metadata: ExchangeMetadata

    hash: string
    nonce?: number | string
    orderHash?: string
    chain?: string
    chainId?: string
}